Add method to declare expected unsubscribed completable counts in tests
We have tests where `Completable` instances get created, but not
subscribed to and this is intentional (for instance, verifying that
certain things do *NOT* happen). For these cases, we need some way to
indicate to the test rule that we expect a certain test to not subscribe
to anything.

This commit adds support for this by adding a meta annotation,
`ExpectUnsubscribed`, which can be used to annotate a test method.